few changes to make code more accessible
